src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ihor Horobets <ihor@horobets.me>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.4.2"
feedparser = "^5.2.1"
pymongo = "^3.10.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
ptpython = "^3.0.1"
feedparser = "^5.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
